Added basic logic around sleeping. beds and bedrolls can now be used. F10 will set up camp. Mana and health regen will use regeneration function. new characters have a hunger value set. widened the uw2 message scroll text display. placeholder functions for killorn crashing, checking for nearby monsters when sleeping, on-map create object traps, added damageplayer function. added playerdat tile state variable

+++ src/base/UWClass.cs
@@ -5,14 +5,14 @@ namespace Underworld
     /// </summary>
     public class UWClass
     {
-        public const int GAME_UWDEMO = 0;//"UW0";
-        public const int GAME_UW1 = 1; // "UW1";
-        public const int GAME_UW2 = 2; // "UW2";
+        public const byte GAME_UWDEMO = 0;//"UW0";
+        public const byte GAME_UW1 = 1; // "UW1";
+        public const byte GAME_UW2 = 2; // "UW2";
 
         /// <summary>
         /// Use to track what game is currently active.
         /// </summary>
-        public static int _RES = 1;  //"UW1";
+        public static byte _RES = 1;  //"UW1";
 
         public static string BasePath  ;    // = "C:\\Games\\UW2\\game\\UW2";
